move to pkg/util
