+++ pkg/util/starlarkutil/stringslice.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package starlarkutils
+package starlarkutil
 
 import (
 	"github.com/cockroachdb/errors"
